// server/src/stream/music/enums.ts
export enum MusicType {
  LOCAL = "LOCAL",
  YT = "YT",
  DB = "DATABASE",
  DB_DOWNLOAD = "DATABASE RE DOWNLOAD"
}

export enum YoutubeSongInfoMessageType {
  NOT_FOUND = "couldn't find any similar songs",
  SONG_TOO_LONG = "your song is too long",
  AGE_RESTRICTED = "cannot play song with age restriction yet",
  PRIVATE = "song is private"
}